fix(covers): UGLY - hardcode jpeg extension to cover cache files
Other than grabbing the Content-Type header (which is cumbersome in the current scope), it is better for now to hardcode the file extension. Most file viewers today can recognize wrong extension and still read, for example, a png file disguised with a .jpg extension;
At least now the cover cache files aren't a bunch of extension-less files anymore.